pyproject.toml
[tool.poetry]
name = "speachrecognitionproject"
version = "0.1.0"
description = ""
authors = ["asultan123 <asultan@aucegypt.edu>"]

[tool.poetry.dependencies]
python = "3.6.7"
SpeechRecognition = "^3.8.1"
pocketsphinx = "^0.1.15"
Flask = "^2.0.1"
Flask-WTF = "^0.15.1"
Flask-Bootstrap4 = "^4.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
